flappyabe completed

+++ flappyabe/javascript/abengers.js
@@ -1,27 +1,23 @@
 class Abengers {
-  constructor(context, arena, num_player, abe, aso) {
+  constructor(context, num_player, abe, aso) {
     this.context = context;
-    this.arena = arena;
 
     const temp = [];
-    temp.push(new Abe(abe, context));
+    temp.push(new Jiminto(abe, context));
     if (num_player == 2) {
-      temp.push(new Aso(aso, context));
+      temp.push(new Jiminto(aso, context));
     }
     this.characters = temp;
-    this.draw();
-  }
-  draw() {
-    this.characters.forEach((character) => {
-      character.draw();
-    });
   }
 
   drop() {
     this.characters.forEach((character) => {
-      character.drop();
+      if (!character.dead) {
+        character.drop();
+      }
     });
   }
+
   jump(p_num) {
     if (this.characters.length == 1) {
       this.characters[0].jump();
@@ -33,26 +29,47 @@ class Abengers {
   getPos(index) {
     return this.characters[index].getPos();
   }
+
+
+  get_status(num) {
+    return this.characters[num].dead;
+  }
+
+  get_gameover() {
+    let bool = true;
+    this.characters.forEach((character) => {
+      //if there is one false, then bool = false;
+      bool = (bool && character.get_status())
+    });
+    return bool;
+  }
+
+  set_status(num, bool) {
+    this.characters[num].dead = bool;
+    this.characters[num].clear();
+  }
 }
 
 class Jiminto {
-  constructor(images, context, x_pos) {
+  constructor(character, context) {
     this.context = context;
     this.img = {
-      rise: images.rise,
-      fall: images.fall,
+      rise: character.rise,
+      fall: character.fall,
     }
     this.pos = {
-      x: x_pos,
+      x: character.x_pos,
       y: 50,
     }
     this.dim = {
-      w: 50,
-      h: 50,
+      w: character.dim,
+      h: character.dim,
     }
     this.rate = 0;
-    this.jumpInitSpeed = -10;
+    this.drop_increment = 0.4;
+    this.jumpInitSpeed = -6;
     this.count_after_rise = 0;
+    this.dead = false;
   }
 
   clear() {
@@ -78,7 +95,6 @@ class Jiminto {
 
   drop() {
     this.count_after_rise++;
-    this.time++;
     this.clear();
     this.pos.y += this.rate;
     if (this.pos.y < 0) {
@@ -88,7 +104,7 @@ class Jiminto {
     else if (this.pos.y > canvas.height - this.dim.h) {
       this.pos.y = canvas.height - this.dim.h;
     }
-    this.rate += 0.5;
+    this.rate += this.drop_increment;
     this.draw();
   }
 
@@ -99,20 +115,12 @@ class Jiminto {
 
   getPos() {
     return {
-      x: this.pos.x,
-      y: this.pos.y,
+      x: this.pos.x + 10,
+      y: (this.pos.y + 5)| 0,
     }
   }
-}
 
-class Abe extends Jiminto {
-  constructor(abe, context) {
-    super(abe, context, 200);
-  }
-}
-
-class Aso extends Jiminto{
-  constructor(aso, context) {
-    super(aso, context, 125);
+  get_status() {
+    return this.dead;
   }
 }
